% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_for_tree.R
\name{convert_for_tree}
\alias{convert_for_tree}
\title{Converts null input to the standard input used for \code{\link[=tree_server]{tree_server()}}}
\usage{
convert_for_tree(object, table)
}
\arguments{
\item{object}{object that gets converted to placeholder missing value if null}

\item{table}{table in \code{\link[=data_model]{data_model()}} to which this object corresponds}
}
\value{
value
}
\description{
Converts a value that is not truthy \code{\link[shiny:isTruthy]{shiny::isTruthy()}} into the  placeholder missing
value for the respective table
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{convert}
\keyword{tree}
