% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_null_date.R
\name{convert_null_date}
\alias{convert_null_date}
\title{Convert the value of a dateInput into an eligible format}
\usage{
convert_null_date(input)
}
\arguments{
\item{input}{input from Shiny}
}
\value{
Date object
}
\description{
dateInputs that return a NULL value can not be written to the database,
but have to be converted into NA
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{NA}
\keyword{NULL}
\keyword{date}
