% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tree.R
\name{create_tree}
\alias{create_tree}
\title{Creates a tree analogous to make_tree, only lowest level is selectable}
\usage{
create_tree(table)
}
\arguments{
\item{table}{name of the table to display, needs to be present in get_dm() environment}
}
\value{
\code{\link[shinytreeview:make_tree]{shinytreeview::make_tree()}} object
}
\description{
Creates a \code{\link[shinytreeview:make_tree]{shinytreeview::make_tree()}} object from a table in the database,
only the lowest level is selectable
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{create_tree}
\keyword{make_tree}
\keyword{shinytreeview}
