% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{dbDataType}
\alias{dbDataType}
\alias{dbDataType,PqConnection,data.frame-method}
\alias{dbDataType,PqConnection,blob-method}
\title{Generate SQL for custom data types}
\usage{
\S4method{dbDataType}{PqConnection,data.frame}(dbObj, obj, ...)

\S4method{dbDataType}{PqConnection,blob}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{Object for determining SQL dialect (e.g. connection)}

\item{obj}{Object for determining SQL type}

\item{...}{Generic consistency}
}
\description{
SQL for data types as described in \code{\link[=blob]{blob()}} can be generated using
\code{\link[DBI:dbDataType]{DBI::dbDataType()}}.
}
