% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_model.R
\name{data_model}
\alias{data_model}
\title{Print data model as list or as dm object}
\usage{
data_model(type)
}
\arguments{
\item{type}{specifies whether to print as "list" or as "dm"}
}
\value{
returns data model as a list of dataframes (when type = "list") or as a dm object (when type = "dm")
}
\description{
Specifies and prints data model as a list of dataframes or as a dm object
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}, Nicolas Bennett, \email{nicolas@cynkra.com}
}
\keyword{data,}
\keyword{model}
