% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_details.R
\docType{data}
\name{db_cons}
\alias{db_cons}
\title{Set database parameters}
\format{
An object of class \code{environment} of length 0.
}
\usage{
db_cons
}
\description{
Sets database parameters
}
\author{
Nicolas Bennett, \email{nicolas@cynkra.com}
}
\keyword{db}
\keyword{details}
