% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_radiology_response.R
\name{display_radiology_response}
\alias{display_radiology_response}
\title{Displays radiology response info in UI}
\usage{
display_radiology_response(input, output, session)
}
\arguments{
\item{input}{input from Shiny}

\item{output}{output to Shiny}

\item{session}{Shiny session}
}
\value{
None
}
\description{
Renders information about radiology response (Recist, Choi, Percist) for the "Radiology" panel in UI
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{radiology}
\keyword{ui}
