% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_column.R
\name{get_column}
\alias{get_column}
\title{Returns a specific column}
\usage{
get_column(table, column)
}
\arguments{
\item{table}{table with the column to select}

\item{column}{column to select}
}
\value{
column from a table in the database
}
\description{
Returns a specific column from a table in the datebase
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{#'}
\keyword{column,}
\keyword{table}
