% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_con.R
\name{get_con}
\alias{get_con}
\title{Get and cache connection}
\usage{
get_con(dbname = db_default(), uid = uid_default(), pwd = pwd_default())
}
\arguments{
\item{dbname}{defaults to db_default()}

\item{uid}{defaults to uid_default()}

\item{pwd}{defaults to pwd_default()}
}
\value{
None
}
\description{
Gets DBI connection
}
\author{
Nicolas Bennett, \email{nicolas@cynkra.com}
}
\keyword{#'}
\keyword{connection}
