% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dm.R
\name{get_dm}
\alias{get_dm}
\title{Get and cache dm object}
\usage{
get_dm(con = get_con())
}
\arguments{
\item{con}{uses get_con to get the current connection to the DB}
}
\value{
dm object
}
\description{
Gets dm object connection and caches it
}
\author{
Nicolas Bennett, \email{nicolas@cynkra.com}
}
\keyword{#'}
\keyword{connection,}
\keyword{dm}
