% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_systemic_drug.R
\name{get_systemic_drug}
\alias{get_systemic_drug}
\title{Returns the ‘drug‘ dataframe with factor variables specified as such}
\usage{
get_systemic_drug()
}
\value{
drug dataframe
}
\description{
Returns the ‘drug‘ dataframe with factor variables specified as such
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{drug}
