% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_systemic_toxicity.R
\name{get_systemic_toxicity}
\alias{get_systemic_toxicity}
\title{Returns the ‘toxicity‘ dataframe with factor variables specified as such}
\usage{
get_systemic_toxicity()
}
\value{
toxicity dataframe
}
\description{
Returns the ‘toxicity‘ dataframe with factor variables specified as such
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{toxicity}
