% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_db.R
\name{init_db}
\alias{init_db}
\title{Initialize the database and load dummy data}
\usage{
init_db(dummy_data = TRUE)
}
\arguments{
\item{dummy_data}{logical whether dummy data should be uploaded to the database, defaults to FALSE}
}
\value{
None
}
\description{
This function removes the current data from the database,
copies the data model to the database and adds tables that are reference tables of foreign key
constraints (tables: insurance, institution, user, toxicity_name, anatomic_region, sarcoma_board_date).
Moreover, dummy data (personal_clinical, cancer_syndrome, sarcoma_board, radiology, pathology, surgery,
amputation, resection, surgery_disciplines) can be uploaded onto the database.
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{init,}
\keyword{initialize}
