% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/institution_input.R
\name{institution_input}
\alias{institution_input}
\title{Create select Input for institution}
\usage{
institution_input(inputId, label)
}
\arguments{
\item{inputId}{inputId used in Shiny's selectInput function}

\item{label}{label used in Shiny's selectInput function}
}
\value{
Shiny selectInput object
}
\description{
Creates a selectInput function for institution
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{institution,}
\keyword{selectInput}
