% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_date.R
\name{is_date}
\alias{is_date}
\title{Checks if object is a date with format "\%Y-\%m-\%d"}
\usage{
is_date(date, date.format = "\%Y-\%m-\%d")
}
\arguments{
\item{date}{object}

\item{date.format}{format of the date object, defaults to "\%Y-\%m-\%d"}
}
\value{
logical
}
\description{
Checks if object is a date with format "\%Y-\%m-\%d"
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{date}
