% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_module_personal.R
\name{personal}
\alias{personal}
\title{UI for the tab "Personal data"}
\usage{
personal()
}
\value{
None
}
\description{
User-interface for the tab "Personal data" which is added to ui.R
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{personal,}
\keyword{ui}
