% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pid_input.R
\name{pid_input}
\alias{pid_input}
\title{Create select Input for PID}
\usage{
pid_input(inputId)
}
\arguments{
\item{inputId}{inputId used in Shiny's selectInput function}
}
\value{
Shiny selectInput object
}
\description{
Creates a selectInput function for PID, chooses the PID column from personal_clinical
using the get_column function
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{PID,}
\keyword{selectInput}
