% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dm.R
\name{remove_dm}
\alias{remove_dm}
\title{Removes data from the database}
\usage{
remove_dm(data, dm, con = get_con(), all = FALSE)
}
\arguments{
\item{data}{dataframes to be removed from the datbase stored as a list}

\item{dm}{dm object}

\item{con}{con object, defaults to get_con() function}

\item{all}{set all = TRUE to delete all data from the database}
}
\value{
None
}
\description{
Removes selected tables or all data from the database using a dm or con object
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}, adapted from Nicolas Bennett, \email{nicolas@cynkra.com}
}
\keyword{#'}
\keyword{remove}
