% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_pathology.R
\name{server_pathology}
\alias{server_pathology}
\title{Server part of the "Pathology" tab}
\usage{
server_pathology(input, output, session, iv)
}
\arguments{
\item{input}{input from Shiny}

\item{output}{output to Shiny}

\item{session}{Shiny session}

\item{iv}{InputValidator from the shinyvalidate package}
}
\value{
None
}
\description{
All server functions for the "Pathology" tab (in page_navbar())
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}, Nicolas Bennett, \email{nicolas@cynkra.com}
}
\keyword{pathology}
\keyword{server}
