% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_radiology.R
\name{server_radiology}
\alias{server_radiology}
\title{Server part of the "Radiology" tab}
\usage{
server_radiology(input, output, session, iv)
}
\arguments{
\item{input}{input from Shiny}

\item{output}{output to Shiny}

\item{session}{Shiny session}

\item{iv}{InputValidator from the shinyvalidate package}
}
\value{
None
}
\description{
All server functions for the "Radiology" tab (in page_navbar())
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}, Nicolas Bennett, \email{nicolas@cynkra.com}
}
\keyword{radiology}
\keyword{server}
