% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_radiotherapy.R
\name{server_radiotherapy}
\alias{server_radiotherapy}
\title{Server part of the "Radiotherapy" tab}
\usage{
server_radiotherapy(input, output, session, iv)
}
\arguments{
\item{input}{input from Shiny}

\item{output}{output to Shiny}

\item{session}{Shiny session}

\item{iv}{InputValidator from the shinyvalidate package}
}
\value{
None
}
\description{
All server functions for the "Radiotherapy" tab (in page_navbar())
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}, Nicolas Bennett, \email{nicolas@cynkra.com}
}
\keyword{radiotherapy}
\keyword{server}
