% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_sb.R
\name{server_sb}
\alias{server_sb}
\title{Server part of the "Sarcoma board" tab}
\usage{
server_sb(input, output, session, iv)
}
\arguments{
\item{input}{input from Shiny}

\item{output}{output to Shiny}

\item{session}{Shiny session}

\item{iv}{InputValidator from the shinyvalidate package}
}
\value{
None
}
\description{
All server functions for the "Sarcoma board" tab (in page_navbar())
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}, Nicolas Bennett, \email{nicolas@cynkra.com}
}
\keyword{sb}
\keyword{server}
