% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_modal.R
\name{tree_modal}
\alias{tree_modal}
\title{Creates a \code{\link[shiny:modalDialog]{shiny::modalDialog()}} with a \code{\link[shinytreeview:treeviewInput]{shinytreeview::treeviewInput()}}}
\usage{
tree_modal(session, input_id, tree, label, selected = NULL, multiple)
}
\arguments{
\item{session}{Shiny session object}

\item{input_id}{input_id for the Shiny inputs in the modal}

\item{tree}{\code{\link[=create_tree]{create_tree()}} or \code{\link[shinytreeview:make_tree]{shinytreeview::make_tree()}} object}

\item{label}{label for Shiny inputs}

\item{selected}{selected value for \code{\link[shinytreeview:treeviewInput]{shinytreeview::treeviewInput()}}, defaults to NULL}

\item{multiple}{logical that indicates if multiple values are allowed for \code{\link[shinytreeview:treeviewInput]{shinytreeview::treeviewInput()}}}
}
\value{
\code{\link[shiny:modalDialog]{shiny::modalDialog()}} object
}
\description{
Creates a \code{\link[shiny:modalDialog]{shiny::modalDialog()}} with a \code{\link[shinytreeview:treeviewInput]{shinytreeview::treeviewInput()}},
a search field, a close treeviewInput button and a dismiss modal button.
The tree is created using the \code{\link[=create_tree]{create_tree()}} function.
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{create_tree}
\keyword{modalDialog}
\keyword{treeviewInput}
