% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_server.R
\name{tree_server}
\alias{tree_server}
\title{Creates a shiny server object that corresponds with \code{\link[=tree_ui]{tree_ui()}}}
\usage{
tree_server(input, output, session, input_id, tree, label, multiple = FALSE)
}
\arguments{
\item{input}{Shiny's input object}

\item{output}{Shiny's output object}

\item{session}{Shiny's session object}

\item{input_id}{input_id for the Shiny inputs in the \code{\link[=tree_modal]{tree_modal()}}}

\item{tree}{\code{\link[=create_tree]{create_tree()}} or \code{\link[shinytreeview:make_tree]{shinytreeview::make_tree()}} object}

\item{label}{label for functions inside the \code{\link[=tree_modal]{tree_modal()}}}

\item{multiple}{logical that indicates if multiple values are allowed for \code{\link[shinytreeview:treeviewInput]{shinytreeview::treeviewInput()}}, defaults to FALSE}
}
\value{
shiny::moduleServer() object
}
\description{
Creates a shiny server object that corresponds with \code{\link[=tree_ui]{tree_ui()}}
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{module}
\keyword{tree_server}
