% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{update_field}
\alias{update_field}
\title{Update Shiny inputs}
\usage{
update_field(fun, id, val, sess)
}
\arguments{
\item{fun}{type of update shiny input function}

\item{id}{inputId of shiny input finction}

\item{val}{value that is used to update}

\item{sess}{Shiny session}
}
\description{
This function calls shiny inputs.
}
