% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_input.R
\name{user_input}
\alias{user_input}
\title{Create select Input for User}
\usage{
user_input(inputId, label, user_discipline, multiple = FALSE)
}
\arguments{
\item{inputId}{inputId used in \code{\link[shiny:selectInput]{shiny::selectInput()}} function}

\item{label}{label used in \code{\link[shiny:selectInput]{shiny::selectInput()}} function}

\item{user_discipline}{discipline to subset by in the user table}

\item{multiple}{logical that indicates if in \code{\link[shiny:selectInput]{shiny::selectInput()}}: multiple = TRUE should be set}
}
\value{
Shiny selectInput object
}
\description{
Creates a \code{\link[shiny:selectInput]{shiny::selectInput()}} function for User by subsetting by desired discipline and
pasting first and last name to a string which is then displayed as choices
}
\author{
Philip Heesen, \email{philip.heesen@uzh.ch}
}
\keyword{selectInput}
\keyword{user,}
