% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dm.R
\name{write_dm}
\alias{write_dm}
\title{Write data to the database}
\usage{
write_dm(data, dm = get_dm())
}
\arguments{
\item{data}{dataframe(s) to write to database saved as a list}

\item{dm}{dm object}
}
\value{
None
}
\description{
Writes dataframes stored as a list to the specified database
}
\author{
Nicolas Bennett, \email{nicolas@cynkra.com}
}
\keyword{write}
